Replace internal implementation with imports
This replaces part of the initial implementation by using functions from
newly imported packages. solid-client exports getWellKnownSolid, which
handles discovering and parsing the .well-known/solid document.
solid-client-vc exports a method to issue a VC, so the consent package
just has to handle the claims in the VC, and not the boilerplate VC
code.