Refactor submodule structure
The functions to request access/consent are now grouped in the
`./request` submodule. The intention is that functions to grant consent
would be grouped in a `./grant` submodule, so that each module may be
imported in the appropriate context. It also makes for smaller files
compared to having everything in one big "consent" module.