Implement approveAccessRequest (#7)
This implements a function to approve a given access request. It
optionally takes in the access request VC or IRI, but enables the person
granting access to override the values from the VC.

Refactoring according to reviews: 
- Rework constants
- Share options type definition
- Fix failing tests after rebase
- Fix linting issue
- Refactor function names to be more accurate: Functions named 'xxConsentRequest' actually accept consent requests as their paramaters, and not access requests which are checked inside the
function to verify that they are indeed consent requests. This lead to a slight refactoring of how the issuance date is managed.