Enable fetching a dataset using a VC
This introduces getSolidDataset, a function wrapping a call to @inrupt/solid-client's getSolidDataset, adding authentication based on a Verifiable Credential representing an access grant.

Note that the function using a VC to build an authenticated fetch isn't exposed yet by solid-client-authn-core, and as such this should not be merged right away.