Replace hasConsent by consentProvided in grants
Consent and Access requests have a `hasConsent` property in their consent subject, while the Consent and Access grants should have a consentProvided property instead. Since it was assumed these two shapes were identical, changing this involved quite a lot of changes across the codebase: the type definitions and type guars had to be updated, and code that handled indistinctively requests and grants now has to check for the type to use the appropriate property.