Redirect to UI with complete VC instead of IRI
This changes the behaviour of redirectToConsentManagementUi: instead of adding the VC IRI as a query parameter, it now adds the VC body, base 64 encoded. The query parameter name changed from `requestVcUrl` to just `requestVc`.

This change is prompted by restrictions on the access control by the consent service: a consent request only references explicitly the requestor, but not the resource owner. Therefore, the resource owner may not dereference the VC IRI to approve/deny the request.