Redirect to UI with complete VC instead of IRI (#33)
This changes the behaviour of redirectToConsentManagementUi: instead of adding the VC IRI as a query parameter, it now adds the VC body, base 64 encoded. The query parameter name changed from `requestVcUrl` to just `requestVc`.

This change is prompted by restrictions on the access control by the consent service: a consent request only references explicitly the requestor, but not the resource owner. Therefore, the resource owner may not dereference the VC IRI to approve/deny the request.

This also adds two options to redirectToConsentManagementUi: one can either specify the resource owner WebID, or directly a fallback UI.The WebID-based consent UI discovery also tries the .well-known of the target resource server.