Discover access issuer from authorization server
This fixes the discovery of the access issuer (formerly known as consent issuer, i.e. a VC service). It used to be discoverable from the `.well-known/solid` document of the Pod server, but it is no longer the case. Instead, it is now part of the `.well-known/uma-configuration` discovery document available at the UMA server contrlling access to resources.