UMA implementation (#53)
UMA enables pushing arbitrary claims in order to get an access token, as described in https://docs.kantarainitiative.org/uma/wg/rec-oauth-uma-grant-2.0.html#uma-grant-type. In our case, the pushed claims are a VC and eventually an ID token in order to get access to a resource. Currently, instead of using an ID token, the request to the token endpoint is authenticated using the access token issued by the OIDC issuer.

This introduces getSolidDataset, getFile, and saveSolidDatasetAt, functions wrapping calls to @inrupt/solid-client's getSolidDataset, adding authentication based on a Verifiable Credential representing an access grant.

Co-authored-by: Nicolas Seydoux <nseydoux@inrupt.com>