Pass VC IRI in redirectToAccessManagementUi
This starts migrating redirectToAccessManagementUi away from passing VC as values, and only passing them as IRIs. For now, both are passed with different query parameter names to avoid conflict. On the next major release, passing the VC as a value can be removed, and the change will be transparent as long as the users have migrated to the new query params accessors.