Override the expiration date of an access request
When approving an access request, one may provide overrides to control exactly what is being granted. In case of an undefined override value, the original value is kept. For expiration dates, it wasn't possible to erase the proposed date and just set the value to nothing. It is now possible using null instead of undefined.