Update jest to 29.x (#388)
* Upgrade jest to latest version
* Change from jose to base64url

Jose ships ESM for the browser, and the transform doesn't work the way it does for solid-client-authn in this context for a reason I cannot figure out. This other dependency should be enough for our purpose.

* Fix jest upgrade issues
* Install ts-node to support jest ts config
* Lower coverage threshold

We are not testing on the auto loading of solid-client-authn-browser session, which creates issues when mocking. This brings our coverage down a bit.