Accept shorter forms for Access Grant status
We currently aren't parsing Access Grant as JSON-LD, but only as JSON. This results in some validations being stricter than necessary: any IRI was expected fully qualified, while it is acceptable to get it in its abbreviated form based on the JSON-LD context.

The current fix is a stopgap, we will address the JSON-LD parsing issue overall in the future.