Normalize access requests on reception
This ensures that Access Request VCs are normalized. For fields where multiple values are possible, if a single value is available, it can be returned as a plain literal instead of an array. That is equivalent in JSON-LD, but we are currently parsing VCs as JSON. To keep our JSON type consistent, a single value is converted back to a one-value array.