Adjust e2e browser based tests
The Dev-2-3 environment introduces a restriction on the clients that can issue an Access Grant. This breaks e2e tests since they're not an allowed client.
The adjustment replaces the use of access grants with access requests, which don't have any restriction and don't need AMC to be involved in the process.