cancelAccessRequest supports revoking an Access Request (#999)
The Dev-2-3 environment introduces a restriction on the clients that can issue an Access Grant. This breaks e2e tests since they're not an allowed client.
The adjustment replaces the use of access grants with access requests, which don't have any restriction and don't need AMC to be involved in the process.

Created an internal function to revoke an access credential (access grant or access request). This function is now used by cancelAccessRequest and revokeAccessGrant. Consequently adapted unit tests.

Downgrade playwright to 1.37.0

---------

Co-authored-by: Aaron Coburn <aaronc@inrupt.com>
Co-authored-by: Pete Edwards <edwardsph@users.noreply.github.com>