Plain JSON VC support (#1027)
Plain JSON VCs are now supported and parsed as JSON-LD as well: defensive statements have been added to check the data type of the VC, as well as a new function to convert the VC into a Dataset if necessary. Additionally, tests have been added to verify this new behaviour.