Add custom fields to Requests and Grants (#1175)
This allows to add custom fields into Access Requests or Access Grants. Initial implementation to read back from them is part of this commit, but expanded in an upcoming commit.

If custom fields overlap with reserved fields, an error is thrown. The `CustomFields` type aims at nudging users in the right direction, by making it explicit the custom field key must be a URL.

The use cases for Access Grants are:
- Issue an Access Grant from an Access Request with custom fields
- Issue an Access Grant overriding some custom fields from the provided
  Access Request
- Issue an Access Grant without an Access Request being present