Add option to verify linked request
When approving an Access Request, one can provide an Access Request so
that the Grant is linked to the Request. By default, this link isn't
verified, and the server doesn't enforce any validation on the linked
Request. The server now supports additional validation, and this flag
allows to opt-in for Access Request validation.

Validation includes:
- Non-revocation of the Request
- Non-expiration of the Request
- The Request must not have been Granted already